[xla:cuda] Delete legacy cuda device allocator
