+++ xla/stream_executor/cuda/BUILD
@@ -836,8 +836,8 @@ cc_library(
         "gpu",
     ],
     deps = [
+        ":cuda_device_allocator",
         ":cuda_status",
-        ":cuda_vmm_allocator",
         "//xla:util",
         "//xla/stream_executor:activate_context",
         "//xla/stream_executor:device_address",
@@ -847,8 +847,6 @@ cc_library(
         "@com_google_absl//absl/log",
         "@com_google_absl//absl/status:statusor",
         "@local_config_cuda//cuda:cuda_headers",
-        "@tsl//tsl/platform:errors",
-        "@tsl//tsl/platform:statusor",
     ],
 )
 
@@ -861,9 +859,9 @@ cc_library(
         "gpu",
     ],
     deps = [
+        ":cuda_device_allocator",
         ":cuda_raw_memory_allocation",
         ":cuda_status",
-        ":cuda_vmm_allocator",
         "//xla:util",
         "//xla/stream_executor:activate_context",
         "//xla/stream_executor:device_address",
@@ -875,7 +873,6 @@ cc_library(
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@local_config_cuda//cuda:cuda_headers",
-        "@tsl//tsl/platform:statusor",
     ],
 )
 
@@ -984,10 +981,10 @@ cc_library(
         "gpu",
     ],
     deps = [
+        ":cuda_device_allocator",
         ":cuda_memory_reservation",
         ":cuda_raw_memory_allocation",
         ":cuda_status",
-        ":cuda_vmm_allocator",
         "//xla/stream_executor:activate_context",
         "//xla/stream_executor:memory_allocation",
         "//xla/stream_executor:memory_reservation",
@@ -996,7 +993,6 @@ cc_library(
         "//xla/stream_executor:stream_executor_h",
         "//xla/stream_executor:vmm_device_address_allocator",
         "//xla/tsl/platform:status_macros",
-        "//xla/tsl/platform:statusor",
         "@com_google_absl//absl/log",
         "@com_google_absl//absl/memory",
         "@com_google_absl//absl/status",
@@ -1208,11 +1204,13 @@ cc_library(
         "//xla/stream_executor:memory_allocation",
         "//xla/stream_executor:memory_allocator",
         "//xla/stream_executor:stream_executor_h",
+        "//xla/tsl/platform:status_macros",
+        "@com_google_absl//absl/base",
+        "@com_google_absl//absl/cleanup",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings:str_format",
         "@local_config_cuda//cuda:cuda_headers",
-        "@tsl//tsl/platform:numbers",
     ],
 )
 
@@ -1230,6 +1228,7 @@ xla_test(
         "//xla/stream_executor:platform_manager",
         "//xla/stream_executor:stream",
         "//xla/stream_executor:stream_executor_h",
+        "//xla/tsl/platform:statusor",
         "@com_google_absl//absl/types:span",
         "@com_google_googletest//:gtest_main",
     ],
@@ -1325,54 +1324,9 @@ xla_test(
     ],
 )
 
-cc_library(
-    name = "cuda_vmm_allocator",
-    srcs = ["cuda_vmm_allocator.cc"],
-    hdrs = ["cuda_vmm_allocator.h"],
-    tags = [
-        "cuda-only",
-        "gpu",
-    ],
-    deps = [
-        ":cuda_status",
-        "//xla:util",
-        "//xla/stream_executor:activate_context",
-        "//xla/stream_executor:device_address",
-        "//xla/stream_executor:memory_allocation",
-        "//xla/stream_executor:memory_allocator",
-        "//xla/stream_executor:stream_executor_h",
-        "//xla/tsl/platform:status_macros",
-        "@com_google_absl//absl/base",
-        "@com_google_absl//absl/cleanup",
-        "@com_google_absl//absl/status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_absl//absl/strings:str_format",
-        "@local_config_cuda//cuda:cuda_headers",
-    ],
-)
-
 xla_test(
-    name = "cuda_vmm_allocator_test",
-    srcs = ["cuda_vmm_allocator_test.cc"],
-    backends = ["gpu"],
-    tags = ["cuda-only"],
-    deps = [
-        ":cuda_platform",
-        ":cuda_vmm_allocator",
-        "//xla/stream_executor:device_address",
-        "//xla/stream_executor:memory_allocation",
-        "//xla/stream_executor:platform",
-        "//xla/stream_executor:platform_manager",
-        "//xla/stream_executor:stream",
-        "//xla/stream_executor:stream_executor_h",
-        "@com_google_absl//absl/types:span",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-xla_test(
-    name = "cuda_vmm_allocator_multigpu_test",
-    srcs = ["cuda_vmm_allocator_multigpu_test.cc"],
+    name = "cuda_device_allocator_multigpu_test",
+    srcs = ["cuda_device_allocator_multigpu_test.cc"],
     backend_tags = {
         "h100": [
             "full",
@@ -1388,15 +1342,16 @@ xla_test(
     tags = ["cuda-only"],
     deps = [
         ":cuda_compute_capability",
+        ":cuda_device_allocator",
         ":cuda_executor",
         ":cuda_platform",
-        ":cuda_vmm_allocator",
         "//xla/stream_executor:device_description",
         "//xla/stream_executor:memory_allocation",
         "//xla/stream_executor:platform",
         "//xla/stream_executor:platform_manager",
         "//xla/stream_executor:stream_executor_h",
         "//xla/stream_executor/gpu:gpu_init",
+        "//xla/tsl/platform:status_macros",
         "@com_google_absl//absl/base:log_severity",
         "@com_google_absl//absl/log:scoped_mock_log",
         "@com_google_absl//absl/status:statusor",
@@ -1657,7 +1612,6 @@ cc_library(
         ":cuda_device_allocator",
         ":cuda_host_allocator",
         ":cuda_kernel",
-        ":cuda_vmm_allocator",
         ":host_callback_registry",
         "//xla/stream_executor:activate_context",
         "//xla/stream_executor:blas",
@@ -1719,7 +1673,6 @@ cc_library(
         ":cuda_timer",
         ":cuda_unified_allocator",
         ":cuda_version_parser",
-        ":cuda_vmm_allocator",
         ":cudnn_api_wrappers",
         ":host_callback_registry",
         ":tma_util",
