+++ xla/stream_executor/cuda/cuda_device_address_vmm_allocator.cc
@@ -31,17 +31,16 @@ limitations under the License.
 #include "xla/tsl/platform/status_macros.h"
 #include "third_party/gpus/cuda/include/cuda.h"
 #include "xla/stream_executor/activate_context.h"
+#include "xla/stream_executor/cuda/cuda_device_allocator.h"
 #include "xla/stream_executor/cuda/cuda_memory_reservation.h"
 #include "xla/stream_executor/cuda/cuda_raw_memory_allocation.h"
 #include "xla/stream_executor/cuda/cuda_status.h"
-#include "xla/stream_executor/cuda/cuda_vmm_allocator.h"
 #include "xla/stream_executor/memory_allocation.h"
 #include "xla/stream_executor/memory_reservation.h"
 #include "xla/stream_executor/platform.h"
 #include "xla/stream_executor/stream.h"
 #include "xla/stream_executor/stream_executor.h"
 #include "xla/stream_executor/vmm_device_address_allocator.h"
-#include "xla/tsl/platform/statusor.h"
 
 namespace stream_executor::gpu {
 
@@ -70,7 +69,7 @@ CudaDeviceAddressVmmAllocator::Create(
     const Platform* platform, double memory_fraction,
     std::optional<int64_t> gpu_system_memory_size,
     absl::Span<const std::pair<StreamExecutor*, Stream*>> devices) {
-  LOG(INFO) << "Using VMM (Virtual Memory Management) allocator.";
+  LOG(INFO) << "Using VMM device-address allocator.";
   std::vector<DeviceConfig> device_configs;
   device_configs.reserve(devices.size());
   for (const auto& [executor, stream] : devices) {
@@ -87,7 +86,7 @@ CudaDeviceAddressVmmAllocator::Create(
     if (gpu_system_memory_size.has_value()) {
       pa_budget = gpu_system_memory_size.value();
     }
-    LOG(INFO) << "VMM allocator pa_budget for device "
+    LOG(INFO) << "VMM device-address allocator pa_budget for device "
               << executor->device_ordinal() << ": " << pa_budget << " bytes.";
     device_configs.push_back({executor, stream, pa_budget});
   }
@@ -140,10 +139,10 @@ absl::Status CudaDeviceAddressVmmAllocator::InitializeDeviceState(
     }
   }
 
-  ASSIGN_OR_RETURN(CudaVmmAllocator::Options vmm_options,
-                   QueryVmmOptions(cu_device));
+  ASSIGN_OR_RETURN(CudaDeviceAllocator::Options device_allocator_options,
+                   QueryDeviceAllocatorOptions(cu_device));
   CUmemAllocationProp alloc_props =
-      BuildVmmAllocationProp(cu_device, vmm_options);
+      BuildVmmAllocationProp(cu_device, device_allocator_options);
   size_t granularity = 0;
   if (auto s = cuda::ToStatus(
           cuMemGetAllocationGranularity(&granularity, &alloc_props,
