+++ xla/stream_executor/cuda/cuda_device_allocator.h
@@ -16,29 +16,60 @@ limitations under the License.
 #ifndef XLA_STREAM_EXECUTOR_CUDA_CUDA_DEVICE_ALLOCATOR_H_
 #define XLA_STREAM_EXECUTOR_CUDA_CUDA_DEVICE_ALLOCATOR_H_
 
+#include <cstddef>
 #include <cstdint>
 #include <memory>
 
 #include "absl/status/statusor.h"
+#include "third_party/gpus/cuda/include/cuda.h"
 #include "xla/stream_executor/memory_allocation.h"
 #include "xla/stream_executor/memory_allocator.h"
 #include "xla/stream_executor/stream_executor.h"
 
 namespace stream_executor::gpu {
 
-// MemoryAllocator that allocates device memory using the CUDA driver API
-// (cuMemAlloc/cuMemFree). This is the lowest-level device memory allocator
-// that directly talks to the CUDA driver without any caching or pooling.
+// CUDA device memory allocator. The current implementation uses CUDA Virtual
+// Memory Management (VMM) APIs (cuMemCreate/cuMemAddressReserve/cuMemMap).
 class CudaDeviceAllocator : public MemoryAllocator {
  public:
+  struct Options {
+    // Minimum alignment for allocations. The actual alignment is the maximum
+    // of this value and the device-reported allocation granularity.
+    size_t alignment = 4096;
+
+    // Whether to enable peer access from all accessible devices.
+    bool enable_peer_access = false;
+
+    // Whether to request POSIX_FILE_DESCRIPTOR handle type.
+    bool enable_posix_fd_handle = true;
+
+    // Whether to request FABRIC handle type.
+    bool enable_fabric_handle = false;
+
+    // Whether to mark allocations as GPUDirect RDMA capable.
+    bool enable_rdma = false;
+  };
+
   explicit CudaDeviceAllocator(StreamExecutor* executor);
+  CudaDeviceAllocator(StreamExecutor* executor, Options options);
 
   absl::StatusOr<std::unique_ptr<MemoryAllocation>> Allocate(
       uint64_t size) final;
 
+  const Options& options() const { return options_; }
+
  private:
   StreamExecutor* executor_;
+  Options options_;
 };
+
+CUmemAllocationProp BuildVmmAllocationProp(
+    CUdevice device, const CudaDeviceAllocator::Options& options);
+
+// Probes device allocator options supported by CUDA VMM; falls back through
+// simpler handle-type combinations if the device rejects the strongest one.
+absl::StatusOr<CudaDeviceAllocator::Options> QueryDeviceAllocatorOptions(
+    CUdevice device);
 
 }  // namespace stream_executor::gpu
