+++ xla/stream_executor/cuda/cuda_device_allocator_test.cc
@@ -28,32 +28,41 @@ limitations under the License.
 #include "xla/stream_executor/platform_manager.h"
 #include "xla/stream_executor/stream.h"
 #include "xla/stream_executor/stream_executor.h"
+#include "xla/tsl/platform/statusor.h"
 
 namespace stream_executor::gpu {
 namespace {
 
-TEST(CudaDeviceAllocatorTest, AllocateAndFree) {
+CudaDeviceAllocator::Options MakeTestOptions(bool enable_rdma) {
+  CudaDeviceAllocator::Options options;
+  options.enable_rdma = enable_rdma;
+  return options;
+}
+
+class CudaDeviceAllocatorTest : public ::testing::TestWithParam<bool> {};
+
+TEST_P(CudaDeviceAllocatorTest, AllocateAndFree) {
   ASSERT_OK_AND_ASSIGN(Platform * platform,
                        PlatformManager::PlatformWithName("CUDA"));
   ASSERT_OK_AND_ASSIGN(StreamExecutor * executor,
                        platform->ExecutorForDevice(0));
 
-  CudaDeviceAllocator allocator(executor);
+  CudaDeviceAllocator allocator(executor, MakeTestOptions(GetParam()));
 
   ASSERT_OK_AND_ASSIGN(std::unique_ptr<MemoryAllocation> allocation,
                        allocator.Allocate(1024));
   ASSERT_NE(allocation, nullptr);
   EXPECT_NE(allocation->address().opaque(), nullptr);
-  EXPECT_EQ(allocation->address().size(), 1024);
+  EXPECT_GE(allocation->address().size(), 1024);
 }
 
-TEST(CudaDeviceAllocatorTest, AllocateZeroBytes) {
+TEST_P(CudaDeviceAllocatorTest, AllocateZeroBytes) {
   ASSERT_OK_AND_ASSIGN(Platform * platform,
                        PlatformManager::PlatformWithName("CUDA"));
   ASSERT_OK_AND_ASSIGN(StreamExecutor * executor,
                        platform->ExecutorForDevice(0));
 
-  CudaDeviceAllocator allocator(executor);
+  CudaDeviceAllocator allocator(executor, MakeTestOptions(GetParam()));
 
   ASSERT_OK_AND_ASSIGN(std::unique_ptr<MemoryAllocation> allocation,
                        allocator.Allocate(0));
@@ -61,7 +70,7 @@ TEST(CudaDeviceAllocatorTest, AllocateZeroBytes) {
   EXPECT_EQ(allocation->address().opaque(), nullptr);
 }
 
-TEST(CudaDeviceAllocatorTest, MemcpyRoundTrip) {
+TEST_P(CudaDeviceAllocatorTest, MemcpyRoundTrip) {
   ASSERT_OK_AND_ASSIGN(Platform * platform,
                        PlatformManager::PlatformWithName("CUDA"));
   ASSERT_OK_AND_ASSIGN(StreamExecutor * executor,
@@ -69,7 +78,7 @@ TEST(CudaDeviceAllocatorTest, MemcpyRoundTrip) {
   ASSERT_OK_AND_ASSIGN(std::unique_ptr<Stream> stream,
                        executor->CreateStream());
 
-  CudaDeviceAllocator allocator(executor);
+  CudaDeviceAllocator allocator(executor, MakeTestOptions(GetParam()));
 
   constexpr int kSize = 1024;
   ASSERT_OK_AND_ASSIGN(std::unique_ptr<MemoryAllocation> allocation,
@@ -81,7 +90,10 @@ TEST(CudaDeviceAllocatorTest, MemcpyRoundTrip) {
     host_src[i] = static_cast<uint8_t>(i);
   }
 
-  DeviceAddress<uint8_t> addr(allocation->address());
+  // Use a DeviceAddress sized to kSize (not the padded allocation size) so
+  // the memcpy transfers exactly the bytes we care about.
+  DeviceAddress<uint8_t> addr(
+      DeviceAddressBase(allocation->address().opaque(), kSize));
   ASSERT_OK(stream->MemcpyH2D(absl::MakeConstSpan(host_src), &addr));
 
   // Copy back from device to host.
@@ -91,6 +103,28 @@ TEST(CudaDeviceAllocatorTest, MemcpyRoundTrip) {
 
   EXPECT_EQ(host_src, host_dst);
 }
+
+TEST_P(CudaDeviceAllocatorTest, PeerAccessEnabled) {
+  ASSERT_OK_AND_ASSIGN(Platform * platform,
+                       PlatformManager::PlatformWithName("CUDA"));
+  ASSERT_OK_AND_ASSIGN(StreamExecutor * executor,
+                       platform->ExecutorForDevice(0));
+
+  CudaDeviceAllocator::Options options = MakeTestOptions(GetParam());
+  options.enable_peer_access = true;
+  CudaDeviceAllocator allocator(executor, options);
+
+  ASSERT_OK_AND_ASSIGN(std::unique_ptr<MemoryAllocation> allocation,
+                       allocator.Allocate(4096));
+  ASSERT_NE(allocation, nullptr);
+  EXPECT_NE(allocation->address().opaque(), nullptr);
+  EXPECT_GE(allocation->address().size(), 4096);
+}
+
+INSTANTIATE_TEST_SUITE_P(RdmaSupport, CudaDeviceAllocatorTest,
+                         ::testing::Bool(), [](const auto& info) {
+                           return info.param ? "RdmaEnabled" : "RdmaDisabled";
+                         });
 
 }  // namespace
 }  // namespace stream_executor::gpu
