+++ xla/stream_executor/cuda/cuda_memory_reservation.cc
@@ -26,14 +26,13 @@ limitations under the License.
 #include "third_party/gpus/cuda/include/cuda.h"
 #include "third_party/gpus/cuda/include/cuda_runtime_api.h"
 #include "xla/stream_executor/activate_context.h"
+#include "xla/stream_executor/cuda/cuda_device_allocator.h"
 #include "xla/stream_executor/cuda/cuda_raw_memory_allocation.h"
 #include "xla/stream_executor/cuda/cuda_status.h"
-#include "xla/stream_executor/cuda/cuda_vmm_allocator.h"
 #include "xla/stream_executor/device_address.h"
 #include "xla/stream_executor/memory_allocation.h"
 #include "xla/stream_executor/stream_executor.h"
 #include "xla/util.h"
-#include "tsl/platform/statusor.h"
 
 namespace stream_executor::gpu {
 
@@ -45,7 +44,8 @@ CudaMemoryReservation::Create(StreamExecutor* executor, uint64_t size) {
   RETURN_IF_ERROR(
       cuda::ToStatus(cuDeviceGet(&device, executor->device_ordinal())));
 
-  ASSIGN_OR_RETURN(CudaVmmAllocator::Options options, QueryVmmOptions(device));
+  ASSIGN_OR_RETURN(CudaDeviceAllocator::Options options,
+                   QueryDeviceAllocatorOptions(device));
   CUmemAllocationProp props = BuildVmmAllocationProp(device, options);
 
   size_t granularity = 0;
